backend/requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
requests==2.28.1
lxml==4.9.1
beautifulsoup4==4.11.1
celery==5.2.7
redis==4.3.4
tqdm==4.64.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Google-Images-Search==1.4.6
pycryptodome==3.14.1
numpy==1.23.5
sentence_transformers==2.2.2
joblib==1.2.0